# Fix struct row comparator's exception on empty structs