Fixes #10603

This PR is to fix a bug of the optimized struct row comparator. For now,  the struct row comparator assumes that structs being compared are non-empty, so it fails when comparing empty structs.