@devavret 2022-04-06
I didn't know there could be empty structs. Does this work if a struct contains 2 struct children, the first of which is empty?