@jrhemstad 2022-04-06
> I didn't know there could be empty structs. Does this work if a struct contains 2 struct children, the first of which is empty?

I'm a bit skeptical of that too, though there doesn't seem to be anything in the Arrow spec that forbids it: https://arrow.apache.org/docs/format/Columnar.html#struct-layout