@jrhemstad 2022-04-06
If empty structs always compare equal, then can't we just strip struct columns without children during the pre-processing step?