@sperlingxx 2022-04-07
> If empty structs always compare equal, then can't we just strip struct columns without children during the pre-processing step?

Because empty structs may carry null masks, which triggers the null comparsion.