@ttnghia 2022-04-07
Rerun tests.