@devavret 2022-04-07
@sperlingxx mind if I push changes to this? I don't think the changes are sufficient. There is probably another corner case here. https://github.com/rapidsai/cudf/pull/10604#issuecomment-1090177986