@sperlingxx 2022-04-07
> SORT_NVBENCH

I added 0 as the depth of struct. Below is the perf result of my local machine: 

```
### [0] NVIDIA TITAN RTX

|     NumRows     | Depth | Nulls | Samples |  CPU Time  | Noise  |  GPU Time  | Noise  |
|-----------------|-------|-------|---------|------------|--------|------------|--------|
|     2^10 = 1024 |     0 |     0 |  11824x |  45.422 us | 11.03% |  42.301 us |  7.77% |
|   2^18 = 262144 |     0 |     0 |   4576x | 112.615 us |  2.82% | 109.586 us |  0.65% |
| 2^26 = 67108864 |     0 |     0 |     44x |  11.388 ms |  0.13% |  11.385 ms |  0.13% |
|     2^10 = 1024 |     1 |     0 |   1630x | 309.766 us |  1.00% | 306.816 us |  0.30% |
|   2^18 = 262144 |     1 |     0 |    499x |   1.006 ms |  0.41% |   1.003 ms |  0.28% |
| 2^26 = 67108864 |     1 |     0 |     11x | 179.348 ms |  0.30% | 179.345 ms |  0.30% |
|     2^10 = 1024 |     8 |     0 |    543x | 924.738 us |  0.37% | 921.886 us |  0.20% |
|   2^18 = 262144 |     8 |     0 |    163x |   3.073 ms |  0.23% |   3.070 ms |  0.21% |
| 2^26 = 67108864 |     8 |     0 |     11x | 589.395 ms |  0.08% | 589.395 ms |  0.08% |
|     2^10 = 1024 |     0 |     1 |   7424x |  70.415 us |  4.55% |  67.393 us |  0.84% |
|   2^18 = 262144 |     0 |     1 |   1716x | 294.552 us |  1.10% | 291.521 us |  0.38% |
| 2^26 = 67108864 |     0 |     1 |     11x |  78.508 ms |  0.09% |  78.505 ms |  0.09% |
|     2^10 = 1024 |     1 |     1 |   1553x | 324.959 us |  1.02% | 322.001 us |  0.43% |
|   2^18 = 262144 |     1 |     1 |   3232x |   1.032 ms | 14.20% |   1.028 ms | 14.20% |
| 2^26 = 67108864 |     1 |     1 |     11x | 180.008 ms |  0.31% | 180.005 ms |  0.31% |
|     2^10 = 1024 |     8 |     1 |   1050x | 940.746 us |  0.60% | 937.648 us |  0.50% |
|   2^18 = 262144 |     8 |     1 |    162x |   3.105 ms |  0.23% |   3.102 ms |  0.20% |
| 2^26 = 67108864 |     8 |     1 |     11x | 588.961 ms |  0.15% | 588.961 ms |  0.15% |
```