@devavret 2022-04-07
> > SORT_NVBENCH
> 
> I added 0 as the depth of struct. Below is the perf result of my local machine:
> 
> ```
> ### [0] NVIDIA TITAN RTX
> 
> |     NumRows     | Depth | Nulls | Samples |  CPU Time  | Noise  |  GPU Time  | Noise  |
> |-----------------|-------|-------|---------|------------|--------|------------|--------|
> | 2^26 = 67108864 |     1 |     0 |     11x | 179.348 ms |  0.30% | 179.345 ms |  0.30% |
> | 2^26 = 67108864 |     8 |     0 |     11x | 589.395 ms |  0.08% | 589.395 ms |  0.08% |
> ```

So a minor perf hit for increasing depth vs old one https://github.com/rapidsai/cudf/pull/10164#issuecomment-1068102865
```
----------------------------------------------------------
Benchmark                                            Time 
----------------------------------------------------------
Sort<false>/unstable/67108864/1/manual_time        174 ms 
Sort<false>/unstable/67108864/8/manual_time        517 ms 
```
Depth 1 time increased from 174 ms to 179 ms and depth 8 increased from 517 ms to 589 ms. Not bad. Correctness is more important.